+++ MovieTicketBooking/src/main/java/com/capg/mtb/repository/IShowRepository.java
@@ -8,13 +8,13 @@ import org.springframework.data.jpa.repository.Query;
 
 import com.capg.mtb.model.Show;
 
-public interface IShowRepository extends JpaRepository<Show, Long> {
+public interface IShowRepository extends JpaRepository<Show, Integer> {
 	
 	@Query(value = "select movie_Id from Shows s where show_Start_Time = 1", nativeQuery = true)
-	List<Long> findByDate(LocalDate date);
+	List<Integer> findByDate(LocalDate date);
 
 	@Query(value = "select * from Shows s where theatre_Id = 1", nativeQuery = true)
-	List<Show> findByTheatreId(long theatreId);
+	List<Show> findByTheatreId(int theatreId);
 
 	@Query(value = "select * from Shows s where show_Start_Time = 1", nativeQuery = true)
 	List<Show> findByLocalDate(LocalDate date);
