+++ MovieTicketBooking/src/main/java/com/capg/mtb/service/IMovieService.java
@@ -3,8 +3,8 @@ package com.capg.mtb.service;
 import java.time.LocalDate;
 import java.util.List;
 
+import com.capg.mtb.exceptions.MovieNotFoundException;
 import com.capg.mtb.model.Movie;
-import com.cg.mtb.exceptions.MovieNotFoundException;
 
 public interface IMovieService  {
 	
@@ -12,13 +12,13 @@ public interface IMovieService  {
 	 
 	 Movie updateMovie(Movie movie);
 	 
-	 void removeMovie(long id) throws MovieNotFoundException;
+	 void removeMovie(int id) throws MovieNotFoundException;
 	 
-	 Movie viewMovie(long id) throws MovieNotFoundException;
+	 Movie viewMovie(int id) throws MovieNotFoundException;
 	 
 	 List<Movie> viewMovieList();
 	 
-	 List<Movie> viewMovieList(long theatreid);
+	 List<Movie> viewMovieList(int theatreid);
 	 
 	 List<Movie> viewMovieList(LocalDate date);
 }
